backend/requirements.txt
fastapi

uvicorn

python-dotenv

supabase

pydantic